+++ app/src/main/java/jt/projects/perfectday/interactors/HolidayInteractorImpl.kt
@@ -2,10 +2,8 @@ package jt.projects.perfectday.interactors
 
 import jt.projects.model.DataModel
 import jt.projects.repository.network.mapper.parseArrayDTOtoDataModel
-import jt.projects.repository.network.retrofit.holiday.HolidayRepository
+import jt.projects.repository.network.holiday.HolidayRepository
 import jt.projects.utils.HOLIDAY_COUNTRY
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.single
 import java.time.LocalDate
 
 class HolidayInteractorImpl(
