+++ repository/src/main/java/jt/projects/repository/network/holiday/HolidayApi.kt
@@ -1,7 +1,7 @@
-package jt.projects.repository.network.retrofit
+package jt.projects.repository.network.holiday
 
 import jt.projects.repository.BuildConfig
-import jt.projects.repository.network.retrofit.holiday.dto.HolidayDTO
+import jt.projects.repository.network.holiday.dto.HolidayDTO
 import retrofit2.http.GET
 import retrofit2.http.Query
